requirements.txt
Django==4.2.2
psycopg2==2.9.5
django-environ==0.9.0
django-allauth==0.51.0
djangorestframework==3.14.0
celery==5.2.7
